+++ sentry-android-core/src/test/java/io/sentry/android/core/performance/AppStartMetricsTest.kt
@@ -154,6 +154,19 @@ class AppStartMetricsTest {
         val metrics = AppStartMetrics.getInstance()
         metrics.registerLifecycleCallbacks(mock<Application>())
 
+        metrics.contentProviderOnCreateTimeSpans.add(
+            TimeSpan().apply {
+                description = "ExampleContentProvider"
+                setStartedAt(1)
+                setStoppedAt(2)
+            }
+        )
+
+        metrics.applicationOnCreateTimeSpan.apply {
+            setStartedAt(3)
+            setStoppedAt(4)
+        }
+
         // when the looper runs
         Shadows.shadowOf(Looper.getMainLooper()).idle()
 
@@ -173,6 +186,8 @@ class AppStartMetricsTest {
         assertTrue(metrics.shouldSendStartMeasurements())
         assertTrue(metrics.appStartTimeSpan.hasStarted())
         assertEquals(now, metrics.appStartTimeSpan.startUptimeMs)
+        assertFalse(metrics.applicationOnCreateTimeSpan.hasStarted())
+        assertTrue(metrics.contentProviderOnCreateTimeSpans.isEmpty())
     }
 
     @Test